Bump project version to 0.0.10 and update dependencies; improve type casting in command list execution and texture upload methods

+++ src/core/FrameGraph/Executor.cpp
@@ -58,5 +58,5 @@ void Executor::Execute(ID3D12Device* device, ID3D12CommandQueue* cmdQueue,
 
   cmdList->Close();
   std::array<ID3D12CommandList*, 1> cmdListArray = {cmdList.Get()};
-  cmdQueue->ExecuteCommandLists(cmdListArray.size(), cmdListArray.data());
+  cmdQueue->ExecuteCommandLists((UINT)cmdListArray.size(), cmdListArray.data());
 }
