+++ src/core/Util.cpp
@@ -234,7 +234,7 @@ _Use_decl_annotations_ HRESULT Util::CreateTexture2DArrayFromMemory(
     return hr;
 
   try {
-    resourceUpload.Upload(res.Get(), 0, subResources, arraySize);
+    resourceUpload.Upload(res.Get(), 0, subResources, (UINT)arraySize);
 
     resourceUpload.Transition(res.Get(), D3D12_RESOURCE_STATE_COPY_DEST,
                               afterState);
